# docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Revendis API",
    "version": "0.2.0",
    "description": "Initial OpenAPI contract for Sprint 1 endpoints.",
    "license": {
      "name": "Proprietary",
      "identifier": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://api.revendis.com",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/sales/orders": {
      "get": {
        "summary": "List sales orders",
        "operationId": "listSalesOrders",
        "tags": [
          "sales"
        ],
        "responses": {
          "200": {
            "description": "Sales list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/sales/checkout": {
      "post": {
        "summary": "Checkout sale",
        "operationId": "checkoutSale",
        "tags": [
          "sales"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaleInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sale confirmed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/sales/orders/{id}/cancel": {
      "post": {
        "summary": "Cancel sale",
        "operationId": "cancelSale",
        "tags": [
          "sales"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sale cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object"
          }
        }
      },
      "ListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
      "ItemResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object"
          }
        }
      },
...
      "SaleItem": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "price": {
            "type": "number"
          }
        }
      },
      "Payment": {
        "type": "object",
        "properties": {
          "method": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          }
        }
      },
      "SaleInput": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "storeId": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SaleItem"
            }
          },
          "discounts": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Payment"
            }
          }
        }
      },
...
    }
  }
}